dataviz/pyproject.toml
[tool.poetry]
name = "dataviz"
version = "0.1.0"
description = "Visualização da camada Gold"
authors = ["Seu Nome"]

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.36.0"
pandas = "^2.2.2"
psycopg2-binary = "^2.9.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
